Make (empty) implementation of the remaining template methods

+++ march_hardware/include/march_hardware/motor_controller/imotioncube/imotioncube.h
@@ -77,7 +77,7 @@ public:
   void goToTargetState(IMotionCubeTargetState target_state);
   virtual void prepareActuation() override;
 
-  uint16_t getSlaveIndex() const override;
+  int getSlaveIndex() const override;
   virtual void reset() override;
 
   /** @brief Override comparison operator */
